+++ 신동현/src/pages/SignUp/SignUpDetail.tsx
@@ -1,9 +1,8 @@
-import { useNavigate, useLocation } from "react-router-dom";
+import { useNavigate, Outlet, useLocation } from "react-router-dom";
 import styled from "styled-components";
 import Button from "@/components/ui/Button/Button";
-import { Add_photo, Kakao, Profile } from "@/assets";
-import Input from "@/components/ui/Input";
-import { useRef, useState } from "react";
+import { Add_photo, Profile } from "@/assets";
+import { createContext, useRef, useState } from "react";
 import Modal from "@/components/ui/Modal/Modal";
 import ActionButton from "@/components/ui/Button/ActionButton";
 import { signUpEmailSchema, signUpSocialSchema } from "@/schema/auth";
@@ -103,56 +102,12 @@ const ProfileChangeContainer = styled.div`
   gap: 20px;
 `;
 
-const InputContainer = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  margin-top: 20px;
-  gap: 25px;
-`;
-
 const ButtonContainer = styled.div`
   width: 100%;
   height: 100%;
   display: flex;
   flex-direction: column;
   margin: 40px 0px 40px 0px;
-`;
-
-const SocialBoxContainer = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-`;
-
-const SocialBoxContext = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding: 12px 16px;
-`;
-
-const SocialBox = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  height: 40px;
-  width: 100%;
-  background: #E6e6e6;
-  border-radius: 4px;
-  font-size: 14px;
-  font-weight: 300;
-  color: #909090;
-`;
-
-const SocialBoxTitle = styled.div`
-  margin-left: 8px;
-  margin-bottom: 8px;
-  font-size: 14px;
-  font-weight: 300;
-  color: #909090;
 `;
 
 const ProfileImageContainer = styled.div`
@@ -178,53 +133,14 @@ const ProfileButton = styled(Button)`
   background-color: transparent;
 `;
 
-const EmailUI = ({ control }: { control: Control<z.infer<typeof signUpEmailSchema>> }) => {
-  return (
-    <InputContainer>
-      <Input title="이메일" type="text" placeholder="이메일" control={control} name="email" value={""} />
-      <Input title="비밀번호" type="password" placeholder="비밀번호" control={control} name="password" value={""} />
-      <Input title="비밀번호 확인" type="password" placeholder="비밀번호 확인" control={control} name="passwordCheck" value={""} />
-      <Input title="이름" type="text" placeholder="이름" control={control} name="name" value={""} />
-      <Input title="생년월일" type="text" placeholder="YYYY-MM-DD" control={control} name="birth" value={""} />
-      <Input title="닉네임" type="text" placeholder="닉네임" control={control} name="nickname" value={""} />
-      <Input title="한 줄 소개" type="text" placeholder="한 줄 소개" control={control} name="bio" value={""} />
-    </InputContainer>
-  )
-}
+export const EmailControlContext = createContext<{ control: Control<z.infer<typeof signUpEmailSchema>> } | null>(null);
+export const SocialControlContext = createContext<{ control: Control<z.infer<typeof signUpSocialSchema>> } | null>(null);
 
-const KaKaoUI = ({ control }: { control: Control<z.infer<typeof signUpSocialSchema>> }) => {
-  const name = localStorage.getItem("name");
 
-  return (
-    <InputContainer>
-      <SocialBoxContainer>
-        <SocialBoxTitle>소셜로그인</SocialBoxTitle>
-        <SocialBox>
-          <SocialBoxContext>
-            <Kakao />
-            카카오 로그인
-          </SocialBoxContext>
-        </SocialBox>
-      </SocialBoxContainer>
-      <Input title="이메일" type="email" placeholder="이메일" control={control} name="email" value={""} />
-      <SocialBoxContainer>
-        <SocialBoxTitle>이름</SocialBoxTitle>
-        <SocialBox>
-          <SocialBoxContext>
-            {name}
-          </SocialBoxContext>
-        </SocialBox>
-      </SocialBoxContainer>
-      <Input title="생년월일" type="text" placeholder="YYYY-MM-DD" control={control} name="birth" value={""} />
-      <Input title="닉네임" type="text" placeholder="닉네임" control={control} name="nickname" value={""} />
-      <Input title="한 줄 소개" type="text" placeholder="한 줄 소개" control={control} name="bio" value={""} />
-    </InputContainer>
-  )
-}
-
-const SignUpDetail = () => {
+const SignUpDetailForm = () => {
   const profilePicture = localStorage.getItem("profilePicture");
   const name = localStorage.getItem("name");
+  const type = useLocation().pathname.split("/")[3];
   const [toast, setToast] = useState<{ message: string; type: "success" | "error" } | null>(null);
   const [profileImageFile, setProfileImageFile] = useState<File | null>(null);
   const [profileImage, setProfileImage] = useState<string | null>(profilePicture);
@@ -232,9 +148,6 @@ const SignUpDetail = () => {
   const [isOpenConfirmModal, setIsOpenConfirmModal] = useState(false);
 
   const navigate = useNavigate();
-
-  const location = useLocation();
-  const type = location.search.split("=")[1];
 
   const { control: controlEmail, handleSubmit: handleSubmitEmail } = useForm<z.infer<typeof signUpEmailSchema>>({
     resolver: zodResolver(signUpEmailSchema),
@@ -353,13 +266,14 @@ const SignUpDetail = () => {
 
   const onCancel = () => {
     closeConfirmModal();
-    navigate("/");
+    navigate("/", { replace: true });
   }
 
   const onConfirm = () => {
     closeConfirmModal();
-    navigate("/?type=login");
+    navigate("/?type=login", { replace: true });
   }
+
 
   return (
     <Wrapper>
@@ -375,14 +289,13 @@ const SignUpDetail = () => {
           <ProfileContent>프로필 사진</ProfileContent>
           <ProfileChangeContainer>
             <input type="file" ref={ImgInputRef} onChange={handleFileChange} style={{ display: 'none' }} />
-
             <ProfileImageContainer>
               <ProfileButton onClick={handleProfileImageChange} icon={profileImage ? <img src={profileImage} alt="profile" /> : <Profile width="90px" height="90px" />} />
             </ProfileImageContainer>
             <Button onClick={handleProfileImageChange} icon={<Add_photo fill="#909090" />} fontSize="12px" width="130px" height="25px" color="#909090" backgroundColor="#FFFFFF" style={{ border: "1px solid #E6E6E6" }}>프로필 사진 추가</Button>
           </ProfileChangeContainer>
         </ProfileContainer >
-        {type === "email" ? <EmailUI control={controlEmail} /> : <KaKaoUI control={controlSocial} />}
+        {type === "email" ? <EmailControlContext.Provider value={{ control: controlEmail }}><Outlet /></EmailControlContext.Provider> : <SocialControlContext.Provider value={{ control: controlSocial }}><Outlet /></SocialControlContext.Provider>}
         <ButtonContainer>
           <ActionButton type="blue" width="100%" height="38px" onClick={type === "email" ? handleSubmitEmail(onSubmit) : handleSubmitSocial(onSubmit)}>회원가입</ActionButton>
         </ButtonContainer>
@@ -393,4 +306,4 @@ const SignUpDetail = () => {
   )
 }
 
-export default SignUpDetail;
+export default SignUpDetailForm;
