+++ 신동현/src/pages/index.tsx
@@ -1,14 +1,16 @@
 import Home from "./Home";
-import SignUp from "./SignUp";
-import SignUpDetail from "./SignUpDetail";
+import SignUp from "./SignUp/SignUp";
+import SignUpDetail from "./SignUp/SignUpDetail";
+import EmailForm from "./SignUp/EmailForm"; 
+import KaKaoForm from "./SignUp/KaKaoForm";
 import Mypage from "./Mypage";
-import KakaoLoginLoading from "./KakaoLoginLoading";
+import KakaoLoginLoading from "./SignUp/KakaoLoginLoading";
 import Detail from "./Detail/Detail";
 import Write from "./Write/Write";
 import WriteForm from "./Write/WriteForm";
 import Update from "./Update/Update";
 import UpdateForm from "./Update/UpdateForm";
 import Root from "./Root";
-import NotFound from "./Not Found";
+import NotFound from "./Error/NotFound";
 
-export { Home, SignUp, SignUpDetail, Mypage, KakaoLoginLoading, Detail, Write, WriteForm, Update, UpdateForm, Root, NotFound };
+export { Home, SignUp, SignUpDetail, EmailForm, KaKaoForm, Mypage, KakaoLoginLoading, Detail, Write, WriteForm, Update, UpdateForm, Root, NotFound };
